YourDGS/requirements.txt
asgiref==3.4.1
astroid==2.3.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
chardet==3.0.4
charset-normalizer==2.0.7
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.6
django-3-jet==1.0.8
django-admin-black==0.0.1
django-ckeditor==6.1.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
google-api-python-client==1.4.1
gunicorn==20.1.0
httplib2==0.20.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.3
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
qrcode==7.3.1
requests==2.22.0
rsa==4.7.2
service-identity==21.1.0
six==1.13.0
sqlparse==0.4.2
sslcommerz-python==0.0.7
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typed-ast==1.4.0
typing-extensions==3.10.0.2
Unipath==1.1
uritemplate==4.1.1
urllib3==1.25.7
whitenoise==5.3.0
wrapt==1.11.2
zope.interface==5.4.0
